Extract changes and where modules
